Extract InlineSVG module from Theme to separate concerns
Move SVG inlining functionality (upload permission, block attribute registration,
render filter, and InlineSVGService) into its own InlineSVG module, keeping Theme
focused on high-level theme setup and configuration.
